+++ include/rt/3d/geometries/model.hpp
@@ -4,6 +4,7 @@
 #include <rt/3d/geometries/geometry.hpp>
 
 #include <memory>
+#include <string>
 
 namespace rt
 {
@@ -14,11 +15,11 @@ class Model final : public Geometry
     Model(const std::string& obj_file_content);
 
     std::span<const Vertex> get_vertices() const override;
-    std::span<const unsigned short> get_indices() const override;
+    std::span<const unsigned int> get_indices() const override;
 
   private:
     std::vector<Vertex> m_vertices;
-    std::vector<unsigned short> m_indices;
+    std::vector<unsigned int> m_indices;
 };
 
 } // namespace rt
